PublishToggle: extract presentational component and PagePublishToggle wrapper
Pulls editor-store reads, draft hiding, and the publish-cascade to
referenced collections out of PublishToggle and into a new
PagePublishToggle wrapper. PublishToggle becomes source-agnostic,
driven by props (isPublic, isSaving, link, onToggle, onRequestUnpublish),
so a future wrapper can reuse it for other entity types. No
user-visible change.
