+++ src/components/PublishToggle.js
@@ -1,72 +1,38 @@
 import { __ } from '@wordpress/i18n';
-import { useSelect, useDispatch } from '@wordpress/data';
-import { store as editorStore } from '@wordpress/editor';
-import { store as blockEditorStore } from '@wordpress/block-editor';
-import { store as coreStore } from '@wordpress/core-data';
 import { Button } from '@wordpress/components';
 import { useCallback, useState } from '@wordpress/element';
 import { globe, lock } from '@wordpress/icons';
 
 /**
- * Walks the block tree and returns unique collectionId values from all
- * cortext/data-view blocks.
+ * Presentational publish toggle. Source-agnostic so it can be reused for
+ * pages (editor store) and collections (core-data) by thin wrappers.
  *
- * @param {Array} blocks Block list to walk.
- * @return {number[]} Unique collection IDs.
+ * @param {Object}    props
+ * @param {boolean}   props.isPublic           Current public state.
+ * @param {boolean}   props.isSaving           Disables the toggle while saving.
+ * @param {?string}   props.link               Public link, when applicable.
+ * @param {Function}  props.onToggle           Called to flip public state.
+ * @param {?Function} props.onRequestUnpublish When set, clicking while public
+ *                                             calls this instead of onToggle so
+ *                                             the wrapper can interpose a
+ *                                             confirmation dialog.
  */
-function getCollectionIds( blocks ) {
-	const ids = new Set();
-	( function walk( list ) {
-		for ( const block of list ) {
-			if (
-				block.name === 'cortext/data-view' &&
-				block.attributes.collectionId
-			) {
-				ids.add( block.attributes.collectionId );
-			}
-			if ( block.innerBlocks?.length ) {
-				walk( block.innerBlocks );
-			}
-		}
-	} )( blocks );
-	return [ ...ids ];
-}
-
-export default function PublishToggle() {
-	const { editPost, savePost } = useDispatch( editorStore );
-	const { saveEntityRecord } = useDispatch( coreStore );
-	const { status, link, isSaving, blocks } = useSelect( ( select ) => {
-		const editor = select( editorStore );
-		return {
-			status: editor.getEditedPostAttribute( 'status' ),
-			link: editor.getEditedPostAttribute( 'link' ),
-			isSaving: editor.isSavingPost(),
-			blocks: select( blockEditorStore ).getBlocks(),
-		};
-	}, [] );
-
+export default function PublishToggle( {
+	isPublic,
+	isSaving,
+	link = null,
+	onToggle,
+	onRequestUnpublish = null,
+} ) {
 	const [ copied, setCopied ] = useState( false );
 
-	const isPublic = status === 'publish';
-
-	const toggle = useCallback( async () => {
-		// Publishing the page also publishes any referenced collections
-		// so their rows become publicly queryable.
-		if ( ! isPublic ) {
-			const collectionIds = getCollectionIds( blocks );
-			await Promise.all(
-				collectionIds.map( ( id ) =>
-					saveEntityRecord( 'postType', 'crtxt_collection', {
-						id,
-						status: 'publish',
-					} )
-				)
-			);
+	const handleClick = useCallback( () => {
+		if ( isPublic && onRequestUnpublish ) {
+			onRequestUnpublish();
+			return;
 		}
-
-		editPost( { status: isPublic ? 'private' : 'publish' } );
-		savePost();
-	}, [ editPost, savePost, saveEntityRecord, isPublic, blocks ] );
+		onToggle();
+	}, [ isPublic, onRequestUnpublish, onToggle ] );
 
 	const copyLink = useCallback( async () => {
 		if ( link ) {
@@ -76,16 +42,11 @@ export default function PublishToggle() {
 		}
 	}, [ link ] );
 
-	// Don't show the toggle for draft pages (no title yet).
-	if ( status === 'draft' ) {
-		return null;
-	}
-
 	return (
 		<div className="cortext-publish-toggle">
 			<Button
 				icon={ isPublic ? globe : lock }
-				onClick={ toggle }
+				onClick={ handleClick }
 				disabled={ isSaving }
 				variant="tertiary"
 				size="compact"
